Make my implementation work with non-ASCII text
The initial commit's implementation was based on Wikipedia's example
implementation. However, this used a fixed alphabet size of 256, apparently for
ASCII characters 0-255. This would not work with non-ASCII text. Now we use
objects to make a 2D representation of the character offset table as described
in Wikipedia's explanation of "The Bad Character Rule".