Restructure code a bit
There is no need for the Boyer-Moore implementation code to be a class
that you must instantiate to use as it has no stateful information.
Additionally, calling it 'grep' is a bit confusing since it doesn't
handle much of the functionality that grep does - primarily searching
files - but that code will end up being implemented elsewhere within the
codebase. At this time that means that the class.extend dependency is
not needed.